+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/grid/GridSynchronizer.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage2.platform.api.grid;
 
+import java.util.List;
 import javax.annotation.Nullable;
 
 import net.minecraft.network.chat.MutableComponent;
@@ -9,6 +10,8 @@ import org.apiguardian.api.API;
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.1.4")
 public interface GridSynchronizer {
     MutableComponent getTitle();
+
+    List<MutableComponent> getHelp();
 
     void synchronizeFromGrid(String text);