+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/grid/StorageChannelTypeSideButtonWidget.java
@@ -5,6 +5,8 @@ import com.refinedmods.refinedstorage2.platform.common.containermenu.grid.Abstra
 import com.refinedmods.refinedstorage2.platform.common.screen.TextureIds;
 import com.refinedmods.refinedstorage2.platform.common.screen.widget.AbstractSideButtonWidget;
 
+import java.util.List;
+
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
 
@@ -13,6 +15,8 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 public class StorageChannelTypeSideButtonWidget extends AbstractSideButtonWidget {
     private static final MutableComponent TITLE = createTranslation("gui", "grid.storage_channel_type");
     private static final MutableComponent SUBTEXT_ALL = createTranslation("gui", "grid.storage_channel_type.all");
+    private static final List<MutableComponent> HELP =
+        List.of(createTranslation("gui", "grid.storage_channel_type.help"));
 
     private final AbstractGridContainerMenu menu;
 
@@ -46,6 +50,11 @@ public class StorageChannelTypeSideButtonWidget extends AbstractSideButtonWidget
             return SUBTEXT_ALL;
         }
         return storageChannelType.getTitle();
+    }
+
+    @Override
+    protected List<MutableComponent> getHelpText() {
+        return HELP;
     }
 
     @Override