+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/widget/FuzzyModeSideButtonWidget.java
@@ -2,6 +2,8 @@ package com.refinedmods.refinedstorage2.platform.common.screen.widget;
 
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
 
+import java.util.List;
+
 import net.minecraft.network.chat.MutableComponent;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
@@ -12,10 +14,15 @@ public class FuzzyModeSideButtonWidget extends AbstractSideButtonWidget {
     private static final MutableComponent SUBTEXT_OFF = createTranslation("gui", "fuzzy_mode.off");
 
     private final ClientProperty<Boolean> property;
+    private final List<MutableComponent> helpOn;
+    private final List<MutableComponent> helpOff;
 
-    public FuzzyModeSideButtonWidget(final ClientProperty<Boolean> property) {
+
+    public FuzzyModeSideButtonWidget(final ClientProperty<Boolean> property, final Type type) {
         super(createPressAction(property));
         this.property = property;
+        this.helpOn = List.of(createTranslation("gui", "fuzzy_mode.on." + type.getHelpTranslationKey()));
+        this.helpOff = List.of(createTranslation("gui", "fuzzy_mode.off." + type.getHelpTranslationKey()));
     }
 
     private static OnPress createPressAction(final ClientProperty<Boolean> property) {
@@ -40,5 +47,21 @@ public class FuzzyModeSideButtonWidget extends AbstractSideButtonWidget {
     @Override
     protected MutableComponent getSubText() {
         return Boolean.TRUE.equals(property.getValue()) ? SUBTEXT_ON : SUBTEXT_OFF;
+    }
+
+    @Override
+    protected List<MutableComponent> getHelpText() {
+        return Boolean.TRUE.equals(property.getValue()) ? helpOn : helpOff;
+    }
+
+    public enum Type {
+        STORAGE,
+        DETECTOR,
+        EXTRACTING_STORAGE_NETWORK,
+        EXTRACTING_SOURCE;
+
+        String getHelpTranslationKey() {
+            return name().toLowerCase() + "_help";
+        }
     }
 }