+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/widget/PrioritySideButtonWidget.java
@@ -3,6 +3,8 @@ package com.refinedmods.refinedstorage2.platform.common.screen.widget;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
 import com.refinedmods.refinedstorage2.platform.common.screen.amount.PriorityScreen;
 
+import java.util.List;
+
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
@@ -13,6 +15,7 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 
 public class PrioritySideButtonWidget extends AbstractSideButtonWidget {
     private static final MutableComponent TITLE = createTranslation("gui", "priority");
+    private static final List<MutableComponent> HELP = List.of(createTranslation("gui", "priority.storage_help"));
 
     private final ClientProperty<Integer> property;
 
@@ -47,5 +50,10 @@ public class PrioritySideButtonWidget extends AbstractSideButtonWidget {
     @Override
     protected MutableComponent getSubText() {
         return Component.literal(String.valueOf(property.getValue()));
+    }
+
+    @Override
+    protected List<MutableComponent> getHelpText() {
+        return HELP;
     }
 }