+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/widget/RedstoneModeSideButtonWidget.java
@@ -3,6 +3,9 @@ package com.refinedmods.refinedstorage2.platform.common.screen.widget;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
 import com.refinedmods.refinedstorage2.platform.common.util.RedstoneMode;
 
+import java.util.List;
+import javax.annotation.Nullable;
+
 import net.minecraft.network.chat.MutableComponent;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
@@ -12,12 +15,26 @@ public class RedstoneModeSideButtonWidget extends AbstractSideButtonWidget {
     private static final MutableComponent SUBTEXT_IGNORE = createTranslation("gui", "redstone_mode.ignore");
     private static final MutableComponent SUBTEXT_HIGH = createTranslation("gui", "redstone_mode.high");
     private static final MutableComponent SUBTEXT_LOW = createTranslation("gui", "redstone_mode.low");
+    private static final MutableComponent HELP_IGNORE = createTranslation("gui", "redstone_mode.ignore.help");
+    private static final MutableComponent HELP_HIGH = createTranslation("gui", "redstone_mode.high.help");
+    private static final MutableComponent HELP_LOW = createTranslation("gui", "redstone_mode.low.help");
 
     private final ClientProperty<RedstoneMode> property;
+    private final List<MutableComponent> helpIgnore;
+    private final List<MutableComponent> helpHigh;
+    private final List<MutableComponent> helpLow;
 
     public RedstoneModeSideButtonWidget(final ClientProperty<RedstoneMode> property) {
+        this(property, null);
+    }
+
+    public RedstoneModeSideButtonWidget(final ClientProperty<RedstoneMode> property,
+                                        @Nullable final MutableComponent extraHelpText) {
         super(createPressAction(property));
         this.property = property;
+        this.helpIgnore = getHelpText(HELP_IGNORE, extraHelpText);
+        this.helpHigh = getHelpText(HELP_HIGH, extraHelpText);
+        this.helpLow = getHelpText(HELP_LOW, extraHelpText);
     }
 
     private static OnPress createPressAction(final ClientProperty<RedstoneMode> property) {
@@ -50,5 +67,22 @@ public class RedstoneModeSideButtonWidget extends AbstractSideButtonWidget {
             case HIGH -> SUBTEXT_HIGH;
             case LOW -> SUBTEXT_LOW;
         };
+    }
+
+    @Override
+    protected List<MutableComponent> getHelpText() {
+        return switch (property.getValue()) {
+            case IGNORE -> helpIgnore;
+            case HIGH -> helpHigh;
+            case LOW -> helpLow;
+        };
+    }
+
+    private static List<MutableComponent> getHelpText(final MutableComponent text,
+                                                      @Nullable final MutableComponent extraText) {
+        if (extraText == null) {
+            return List.of(text);
+        }
+        return List.of(text, extraText);
     }
 }