+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/integration/recipemod/rei/ReiGridSynchronizer.java
@@ -2,6 +2,7 @@ package com.refinedmods.refinedstorage2.platform.fabric.integration.recipemod.re
 
 import com.refinedmods.refinedstorage2.platform.common.internal.grid.AbstractGridSynchronizer;
 
+import java.util.List;
 import javax.annotation.Nullable;
 
 import net.minecraft.network.chat.MutableComponent;
@@ -11,6 +12,9 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 public class ReiGridSynchronizer extends AbstractGridSynchronizer {
     private static final MutableComponent TITLE = createTranslation("gui", "grid.synchronizer.rei");
     private static final MutableComponent TITLE_TWO_WAY = createTranslation("gui", "grid.synchronizer.rei.two_way");
+    private static final List<MutableComponent> HELP = List.of(createTranslation("gui", "grid.synchronizer.rei.help"));
+    private static final List<MutableComponent> HELP_TWO_WAY =
+        List.of(createTranslation("gui", "grid.synchronizer.rei.two_way.help"));
 
     private final ReiProxy reiProxy;
     private final boolean twoWay;
@@ -23,6 +27,11 @@ public class ReiGridSynchronizer extends AbstractGridSynchronizer {
     @Override
     public MutableComponent getTitle() {
         return twoWay ? TITLE_TWO_WAY : TITLE;
+    }
+
+    @Override
+    public List<MutableComponent> getHelp() {
+        return twoWay ? HELP_TWO_WAY : HELP;
     }
 
     @Override